// nestjs-base/src/config/typeorm-config/types.ts
export interface ITypeormConfig {
	GLOBAL_TYPEORM_TYPE: 'mysql';
	GLOBAL_TYPEORM_HOST: string;
	GLOBAL_TYPEORM_PORT: number;
	GLOBAL_TYPEORM_USERNAME: string;
	GLOBAL_TYPEORM_PASSWORD: string;
	GLOBAL_TYPEORM_DATABASE: string;
	GLOBAL_TYPEORM_SYNCHRONIZE: boolean;
	GLOBAL_TYPEORM_LOGGING: boolean;
}